+++ block/CHANGELOG.md
@@ -1,9 +1,9 @@
-# @platforma-open/milaboratories.wip-cell-browser
+# @platforma-open/milaboratories.cell-browser-2
 
 ## 1.0.1
 
 ### Patch Changes
 
 - Updated dependencies [9e7c9c9]
-  - @platforma-open/milaboratories.wip-cell-browser.model@1.0.1
-  - @platforma-open/milaboratories.wip-cell-browser.ui@1.0.1
+  - @platforma-open/milaboratories.cell-browser-2.model@1.0.1
+  - @platforma-open/milaboratories.cell-browser-2.ui@1.0.1