+++ test/CHANGELOG.md
@@ -1,8 +1,8 @@
-# @platforma-open/milaboratories.wip-cell-browser.test
+# @platforma-open/milaboratories.cell-browser-2.test
 
 ## 1.0.1
 
 ### Patch Changes
 
 - Updated dependencies [9e7c9c9]
-  - @platforma-open/milaboratories.wip-cell-browser.model@1.0.1
+  - @platforma-open/milaboratories.cell-browser-2.model@1.0.1