+++ ui/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @platforma-open/milaboratories.wip-cell-browser.ui
+# @platforma-open/milaboratories.cell-browser-2.ui
 
 ## 1.0.1
 
@@ -6,4 +6,4 @@
 
 - 9e7c9c9: Advanced annotations filters
 - Updated dependencies [9e7c9c9]
-  - @platforma-open/milaboratories.wip-cell-browser.model@1.0.1
+  - @platforma-open/milaboratories.cell-browser-2.model@1.0.1